requirements.txt
aiogram
aiomysql # Асинхронный драйвер для MySQL
python-dotenv # Для загрузки переменных окружения
cryptography
